Создаёт js модульность
